requirements.txt
pytest==7.0.1
selenium==4.1.3
allure-pytest==2.8.32
allure-python-commons==2.8.32
webdriver-manager==3.5.3
requests==2.28.1

